Implemented support for stream table definitions (formerly subscriptions). This required a complete reworking of the type system which is now unified around the UniversalTableBuilder and RelDataType.

+++ sqrl-core/src/main/java/ai/datasqrl/schema/converters/SourceRecord2RowMapper.java
@@ -58,7 +58,7 @@ public class SourceRecord2RowMapper<R> implements Function<SourceRecord.Named, R
                                 Object[] cols = constructRows(item, subType);
                                 //Add index
                                 cols = extendCols(cols,1);
-                                cols[0] = Long.valueOf(idx);
+                                cols[0] = Integer.valueOf(idx);
                                 rows[idx] = rowConstructor.createNestedRow(cols);
                                 idx++;
                             }
